Split file indexer ignore rules
